优化 Event Function Fixes #11
