+++ lib/Amateras.js
@@ -53,7 +53,6 @@ class Amateras {
             console.time('| System Initialized');
             yield this.guilds.init();
             console.timeEnd('| System Initialized');
-            this.interactionCreate();
             this.eventHandler();
             this.setTimer();
             this.me = yield this.players.fetch(this.id);
@@ -100,127 +99,6 @@ class Amateras {
                 }
             });
         });
-    }
-    interactionCreate() {
-        // Client Interaction Command
-        this.client.on('interactionCreate', (interaction) => __awaiter(this, void 0, void 0, function* () {
-            var _a, _b;
-            let consoleText = `command received: ${interaction.user.username} - `;
-            if (interaction.isCommand()) { // If slash command message sent
-                // Check command file exist
-                if (fs_1.default.existsSync(`./commands/${interaction.commandName}.js`)) {
-                    // Import command function
-                    const commandFn = require(`../commands/${interaction.commandName}.js`);
-                    // Call command Function
-                    commandFn.default(interaction, this);
-                    consoleText += `${interaction.commandName}\n${interaction.options.data}`;
-                }
-                else {
-                    // If command file not exist
-                    console.error('Command not exist. Function file not found. (Amateras.js)');
-                    return;
-                }
-            }
-            else if (interaction.isSelectMenu()) { // If menu selected
-                const flags = interaction.message.flags;
-                if (flags.toArray().includes('EPHEMERAL'))
-                    return;
-                const msg = (yield ((_a = this.messages) === null || _a === void 0 ? void 0 : _a.fetch(interaction.message.id)));
-                if (!msg) {
-                    console.error(`Msg "${interaction.id}" not found. (Amateras.js)`);
-                    return;
-                }
-                let SelectFn = '';
-                if (!msg.actions) {
-                    console.error(`Msg "${interaction.id}" actions is ${msg.actions}. (Amateras.js)`);
-                    return;
-                }
-                for (const action of msg.actions) {
-                    for (const element of action) {
-                        if (element.type === 'SELECT_MENU' && Object.keys(element.options).includes(interaction.values[0])) {
-                            if (!element.options[interaction.values[0]]) {
-                                console.error(`Msg "${msg.id}" Select menu "${element.customId}" function is ${element.options[interaction.values[0]]} (Amateras.js)`);
-                                return;
-                            }
-                            SelectFn = element.options[interaction.values[0]];
-                        }
-                    }
-                }
-                if (!SelectFn) {
-                    //console.error(`Select menu "${interaction.customId}" not found in msg "${msg.id}" or type is uncorrect. (Amateras.js)`)
-                    return;
-                }
-                consoleText += `${interaction.customId}\n${interaction.values}`;
-                // Check function file exist
-                if (fs_1.default.existsSync(`./reacts/${SelectFn}.js`)) {
-                    // Import function
-                    const fn = require(`../reacts/${SelectFn}.js`);
-                    // Call Function
-                    fn.default(interaction, this, msg.get);
-                }
-                else {
-                    // If Function not exist
-                    console.error(`Select menu "${interaction.customId}" function file not found. (Amateras.js)`);
-                    return;
-                }
-            }
-            else if (interaction.isContextMenu()) { // If context menu clicked
-                if (fs_1.default.existsSync(`./context/${interaction.commandName}.js`)) {
-                    // Import command function
-                    const contextFn = require(`../context/${interaction.commandName}.js`);
-                    // Call command Function
-                    contextFn.default(interaction, this);
-                    consoleText += `${interaction.commandName}`;
-                }
-            }
-            else if (interaction.isButton()) { // If button clicked
-                const flags = interaction.message.flags;
-                if (flags.toArray().includes('EPHEMERAL'))
-                    return;
-                const msg = (yield ((_b = this.messages) === null || _b === void 0 ? void 0 : _b.fetch(interaction.message.id)));
-                if (!msg) {
-                    console.error(`Msg "${interaction.id}" not found. (Amateras.js)`);
-                    return;
-                }
-                let buttonFn = '';
-                if (!msg.actions) {
-                    console.error(`Msg "${interaction.id}" actions is ${msg.actions}. (Amateras.js)`);
-                    return;
-                }
-                const customId = interaction.customId.split('#')[1];
-                for (const action of msg.actions) {
-                    for (const element of action) {
-                        if (element.customId === customId && element.type === 'BUTTON') {
-                            if (!element.fn) {
-                                console.error(`Msg "${msg.id}" button "${element.customId}" function is ${element.fn} (Amateras.js)`);
-                                return;
-                            }
-                            buttonFn = element.fn;
-                        }
-                    }
-                }
-                if (!buttonFn) {
-                    console.error(`Button "${customId}" not found in msg "${msg.id}" or type is uncorrect. (Amateras.js)`);
-                    return;
-                }
-                consoleText += `${customId}`;
-                // Check function file exist
-                if (fs_1.default.existsSync(`./reacts/${buttonFn}.js`)) {
-                    // Import function
-                    const fn = require(`../reacts/${buttonFn}.js`);
-                    // Call Function
-                    fn.default(interaction, this);
-                }
-                else {
-                    // If Function not exist
-                    console.error(`Button "${interaction.customId}" function file not found. (Amateras.js)`);
-                    return;
-                }
-            }
-            else
-                return;
-            console.log(true, consoleText);
-        }));
     }
     eventHandler() {
         const eventFiles = fs_1.default.readdirSync('./events').filter(file => file.endsWith('.js'));
